+++ src/main/java/core_entities/game_parts/board/BoardFactory.java
@@ -1,4 +1,4 @@
-package core_entities.game_parts;
+package core_entities.game_parts.board;
 
 public interface BoardFactory {
     Board create(String [][] multipliers);
