+++ src/main/java/core_entities/game_parts/board/DefaultBoardFactory.java
@@ -1,4 +1,4 @@
-package core_entities.game_parts;
+package core_entities.game_parts.board;
 
 public class DefaultBoardFactory implements BoardFactory{
     public Board create(String [][] multipliers) {
