+++ src/main/java/core_entities/game_parts/dictionary/DictionarySearch.java
@@ -1,4 +1,4 @@
-package core_entities.game_parts;
+package core_entities.game_parts.dictionary;
 
 import java.util.ArrayList;
