+++ src/main/java/use_cases/place_word_refill_user_story/boardChecker.java
@@ -1,6 +1,6 @@
 package use_cases.place_word_refill_user_story;
 
-import core_entities.game_parts.Board;
+import core_entities.game_parts.board.Board;
 import core_entities.game_parts.Coordinate;
 import core_entities.game_parts.GameState;
